+++ cmd/neofs-ir/defaults.go
@@ -74,7 +74,6 @@ func defaultConfiguration(cfg *viper.Viper) {
 	cfg.SetDefault("workers.container", "10")
 	cfg.SetDefault("workers.alphabet", "10")
 	cfg.SetDefault("workers.reputation", "10")
-	cfg.SetDefault("workers.subnet", "10")
 
 	cfg.SetDefault("netmap_cleaner.enabled", true)
 	cfg.SetDefault("netmap_cleaner.threshold", 3)