+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changelog for NeoFS Node
 ### Removed
 - Deprecated `morph.rpc_endpoint` SN and `morph.endpoint.client` IR config sections (#2400)
 - `neofs_node_object_epoch` metric for IR and SN (#2347)
+- Subnets support (#2411)
 
 ### Changed
 - CLI `--timeout` flag configures whole execution timeout from now (#2124)
@@ -32,6 +33,11 @@ have been removed. Use `morph.endpoints` for both instead.
 Deprecated `neofs_node_object_epoch` metric for IR and SN (the same for both)
 has been removed. Use `neofs_node_state_epoch` for SN and `neofs_ir_state_epoch`
 for IR instead.
+Subnets support has been removed:
+- IR's `workers.subnet` and `contracts.subnet` configs are not used anymore.
+- SN's `node.subnet` config section is not used anymore.
+- `neoofs-amd morph` does not have `subnet` subcommand anymore.
+- `neofs-cli container create` does not have `--subnet` flag anymore.
 
 ## [0.37.0] - 2023-06-15 - Sogado